chore: load sample code from filesystem during build
