+++ package.json
@@ -23,6 +23,7 @@
     "eslint": "8.23.1",
     "eslint-config-next": "12.3.0",
     "postcss": "^8.4.16",
+    "raw-loader": "^4.0.2",
     "tailwindcss": "^3.1.8"
   }
 }
